expandAndCollapseHelper 함수 생성
